+++ winter/messaging/__init__.py
@@ -1,5 +1,5 @@
 from .event import Event
 from .event_handler import event_handler
-from .event_handler_registry import EventHandlerRegistry
+from .event_subscription_registry import EventSubscriptionRegistry
 from .event_publisher import EventPublisher
 from .simple_event_publisher import SimpleEventPublisher