+++ winter/messaging/event_subscription.py
@@ -4,28 +4,30 @@ from typing import Callable
 from typing import Tuple
 from typing import Type
 
+from winter.core import ComponentMethod
 from winter.core.utils.typing import get_generic_args
 from winter.core.utils.typing import is_iterable_type
 from winter.core.utils.typing import is_union
-from .domain_event import DomainEvent
+from winter.messaging import Event
 
 
 @dataclass(frozen=True)
-class DomainEventSubscription:
-    event_filter: Tuple[Type[DomainEvent]]
+class EventSubscription:
+    event_filter: Tuple[Type[Event]]
     collection: bool
-    handler_class: Type
-    handler_method: Callable
+    handler_method: ComponentMethod
 
     @staticmethod
-    def create(handler_class, handler_method):
-        func_spec = inspect.getfullargspec(handler_method)
-        arg_type = func_spec.annotations[func_spec.args[1]]
+    def create(handler_method: ComponentMethod):
+        arg_type = handler_method.arguments[0].type_
         collection = is_iterable_type(arg_type)
+
         if collection:
             arg_type = get_generic_args(arg_type)[0]
+
         if is_union(arg_type):
             domain_event_classes = tuple(get_generic_args(arg_type))
         else:
             domain_event_classes = (arg_type, )
-        return DomainEventSubscription(domain_event_classes, collection, handler_class, handler_method)
+
+        return EventSubscription(domain_event_classes, collection, handler_method)